Fix clang warnings in HTTP single client

+++ include/solana_client/rpc_single_ws_request_client.hpp
@@ -16,7 +16,7 @@ typedef struct {
 typedef struct {
 	Dictionary request;
 	Dictionary parsed_url;
-	float timeout;
+	double timeout;
 	int request_identifier;
 	Callable callback;
 	Callable confirmation_callback;
